+++ pennylane/ops/qubit.py
@@ -1722,10 +1722,15 @@ class SingleExcitation(Operation):
 
     .. code-block::
 
+        dev = qml.device('default.qubit', wires=2)
+
         @qml.qnode(dev)
         def circuit(phi):
             qml.PauliX(wires=0)
             qml.SingleExcitation(phi, wires=[0, 1])
+            return qml.state()
+
+        circuit(0.1)
     """
 
     num_params = 1
@@ -2026,7 +2031,7 @@ class MultiControlledX(ControlledQubitUnitary):
     ``ControlledQubitUnitary``, but there is no need to specify a matrix
     argument:
 
-    >>> qml.MultiControlledX(control_wires=[0, 1, 2, 3], wires=4, control_values='1110'])
+    >>> qml.MultiControlledX(control_wires=[0, 1, 2, 3], wires=4, control_values='1110')
 
     """
     num_params = 1
@@ -2119,12 +2124,15 @@ class QFT(Operation):
 
         wires = 3
 
+        dev = qml.device('default.qubit',wires=wires)
+
         @qml.qnode(dev)
         def circuit_qft(basis_state):
             qml.BasisState(basis_state, wires=range(wires))
             qml.QFT(wires=range(wires))
             return qml.state()
 
+        circuit_qft([1.0, 0.0, 0.0])
     """
     num_params = 0
     num_wires = AnyWires
@@ -2217,11 +2225,16 @@ class DoubleExcitation(Operation):
 
     .. code-block::
 
+        dev = qml.device('default.qubit', wires=4)
+
         @qml.qnode(dev)
         def circuit(phi):
             qml.PauliX(wires=0)
             qml.PauliX(wires=1)
             qml.DoubleExcitation(phi, wires=[0, 1, 2, 3])
+            return qml.state()
+
+        circuit(0.1)
     """
 
     num_params = 1